+++ AdvancedTodoList.IntegrationTests/Factories/BusinessLogicWebApplicationFactory.cs
@@ -1,4 +1,5 @@
 ﻿using AdvancedTodoList.Core.Models.TodoLists;
+using AdvancedTodoList.Core.Repositories;
 using AdvancedTodoList.Core.Services.Repositories;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Mvc.Testing;
