+++ AdvancedTodoList/Program.cs
@@ -6,9 +6,9 @@ using AdvancedTodoList.Core.Services.Auth;
 using AdvancedTodoList.Core.Services.Repositories;
 using AdvancedTodoList.Core.Validation;
 using AdvancedTodoList.Infrastructure.Data;
+using AdvancedTodoList.Infrastructure.Repositories;
 using AdvancedTodoList.Infrastructure.Services;
 using AdvancedTodoList.Infrastructure.Services.Auth;
-using AdvancedTodoList.Infrastructure.Services.Repositories;
 using FluentValidation;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.EntityFrameworkCore;
